fix: 修复编译错误（AAPT/import 路径/类型推断）
- AndroidManifest: <uses-native-library> 移入 <application> 内（AAPT 要求）
- VideoConfig: 0xFFFFFFFE 加 .toInt()，避免 Long 字面量与 Int 做 and 得 Long
- NodeRuntime: 补 import android.graphics.Bitmap
- ChatScreen: rememberSaveable 改用 androidx.compose.runtime.saveable 包路径
- SceneRenderer: resume 改用 stdlib kotlin.coroutines.resume，去掉空 onCancellation 回调

+++ app/src/main/java/com/myagent/app/multimodal/VideoConfig.kt
@@ -34,8 +34,8 @@ data class VideoConfig(
       val parts = raw.split(",").map { it.trim() }
       return try {
         VideoConfig(
-          width = parts.getOrElse(0) { "854" }.toInt().coerceIn(320, 1920) and 0xFFFFFFFE,
-          height = parts.getOrElse(1) { "480" }.toInt().coerceIn(240, 1080) and 0xFFFFFFFE,
+          width = parts.getOrElse(0) { "854" }.toInt().coerceIn(320, 1920) and 0xFFFFFFFE.toInt(),
+          height = parts.getOrElse(1) { "480" }.toInt().coerceIn(240, 1080) and 0xFFFFFFFE.toInt(),
           fps = parts.getOrElse(2) { "24" }.toInt().coerceIn(12, 60),
           maxDuration = parts.getOrElse(3) { "5" }.toInt().coerceIn(DURATION_MIN, DURATION_MAX),
         )
